+++ jumpapp/assets/js/src/classes/Weather.js
@@ -1,34 +1,27 @@
 export default class Weather {
 
     /**
-     * Reposible for retrieveing weather data from OWM and doing
+     * Responsible for retrieveing weather data from OWM and doing
      * stuff with it.
      *
-     * @param {string} owmapikey OWM API key.
      * @param {string} latlong Comma separated string representing a lattitude and longitude.
-     * @param {boolean} metrictemp Are temperature units in metric or imperial.
      */
-    constructor(owmapikey, latlong, metrictemp) {
-        this.owmapiurlbase = 'https://api.openweathermap.org/data/2.5/weather';
-        this.owmapikey = owmapikey;
-        this.latlong = latlong;
-        this.metrictemp = metrictemp
+    constructor(eventemitter) {
+        this.eventemitter = eventemitter;
     }
 
     /**
-     * Make an async request to the OWM API, parse and return the response.
-     *
-     * @returns {Promise} Containing parsed OWM data.
+     * Make an async request to the weather API, parse and return the response.
      */
-    async fetch_owm_data() {
-        const url = this.owmapiurlbase
-                    +'?units=' + (this.metrictemp ? 'metric' : 'imperial')
-                    +'&lat=' + this.latlong[0]
-                    +'&lon=' + this.latlong[1]
-                    +'&appid=' + this.owmapikey;
-
-        // Get some data from the open weather map api...
-        const promise = await fetch(url)
+    fetch_owm_data(latlong) {
+        // If we are provided with a latlong then the user must have cliecked on the location
+        // button at some point, so let's use this in the api url...
+        let apiurl = '/api/weatherdata.php';
+        if (latlong.length) {
+            apiurl += ('?lat=' + latlong[0] + '&lon=' + latlong[1]);
+        }
+        // Get some data from the weather api...
+        fetch(apiurl)
         .then(response => response.json())
         .then(data => {
             if (data.cod === 401) {
@@ -39,16 +32,15 @@ export default class Weather {
             if (data.dt > data.sys.sunrise && data.dt < data.sys.sunset) {
                 daynightvariant = 'day'
             }
-            return {
+            this.eventemitter.emit('weather-loaded', {
                 locationcode: data.id,
                 locationname: data.name,
-                temp: Math.ceil(data.main.temp) + '&deg;' + (this.metrictemp ? 'C' : 'F'),
+                temp: Math.ceil(data.main.temp) + '&deg;' + (JUMP.metrictemp ? 'C' : 'F'),
                 description: data.weather[0].main,
                 iconclass: 'wi-owm-' + daynightvariant + '-' + data.weather[0].id,
                 timezoneshift: data.timezone
-            };
+            });
         })
-        return promise;
     }
 
 }