+++ jumpapp/background-css.php
@@ -10,7 +10,10 @@
 require __DIR__ .'/vendor/autoload.php';
 
 $config = new Jump\Config();
-$backgroundfile = (new Jump\Background($config))->get_random_background_file();
+$backgroundimgfile = (new Jump\Background($config))->get_random_background_file();
+
+$blur = floor((int)$config->get('bgblur', false) / 100 * 15);
+$brightness = (int)$config->get('bgbright', false) ? (int)$config->get('bgbright', false) / 100 : 1;
 
 header('Content-Type: text/css');
-echo '.background {background-image: url("'.$backgroundfile.'");}';
+echo '.background {background-image: url("'.$backgroundimgfile.'");filter: brightness('.$brightness.') blur('.$blur.'px);}';