+++ jumpapp/classes/Exceptions/TagNotFoundException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Jump\Exceptions;
+
+class TagNotFoundException extends \Exception {}