+++ jumpapp/classes/Site.php
@@ -2,6 +2,8 @@
 
 namespace Jump;
 
+use stdClass;
+
 /**
  * Parse the data required to represent a site and provide method for generating
  * and/or retrieving the site's icon.
@@ -11,43 +13,49 @@ namespace Jump;
  */
 class Site {
 
-    private Config $config;
     public string $name;
     public bool $nofollow;
-    public string $icon;
+    public ?string $iconname;
     public string $url;
+    public array $tags = ['home'];
 
-    public function __construct(Config $config, array $sitearray, array $default) {
-        $this->config = $config;
-        $this->defaults = $default;
+    /**
+     * Parse the data required to represent a site and provide method for generating
+     * and/or retrieving the site's icon.
+     *
+     * @param Config $config A Jump Config() object.
+     * @param array $sitearray Array of options for this site from sites.json.
+     * @param array $defaults Array of default values for this site to use, defined in sites.json.
+     */
+    public function __construct(private Config $config, array $sitearray, private array $defaults) {
         if (!isset($sitearray['name'], $sitearray['url'])) {
             throw new \Exception('The array passed to Site() must contain the keys "name" and "url"!');
         }
         $this->name = $sitearray['name'];
         $this->url = $sitearray['url'];
         $this->nofollow = isset($sitearray['nofollow']) ? $sitearray['nofollow'] : (isset($this->defaults['nofollow']) ? $this->defaults['nofollow'] : false);
-        $this->icon = isset($sitearray['icon']) ? $this->get_favicon_datauri($sitearray['icon']) : $this->get_favicon_datauri();
+        $this->iconname = $sitearray['icon'] ?? null;
+        $this->tags = $sitearray['tags'] ?? $this->tags;
     }
 
     /**
-     * Return a data uri for a given icon, or a site's favicon if an icon
-     * is not provided.
+     * Return an object containing mimetype and raw image data, or a site's
+     * favicon if an icon is not provided in sites.json.
      *
-     * @param string|null $icon File name of a given icon to retrieve.
-     * @return string Base 64 encoded datauri for the icon image.
+     * @return object Containing mimetype and raw image data.
      */
-    public function get_favicon_datauri(?string $icon = null): string {
+    public function get_favicon_image_data(): object {
         // Use the applications own default icon unless one is supplied via the sites.json file.
         $defaulticon = $this->config->get('defaulticonpath');
         if (isset($this->defaults['icon'])) {
             $defaulticon = $this->config->get('sitesdir').'/icons/'.$this->defaults['icon'];
         }
         // Did we have a supplied icon or are we going to try retrieving the favicon?
-        if ($icon === null) {
+        if ($this->iconname === null) {
             // Go get the favicon, if there isnt one then use the default icon.
             $favicon = new \Favicon\Favicon();
             $favicon->cache([
-                'dir' => $this->config->get('cachedir').'/icons/',
+                'dir' => $this->config->get('cachedir').'/icons',
                 'timeout' => 86400
             ]);
             $rawimage = $favicon->get($this->url, \Favicon\FaviconDLType::RAW_IMAGE);
@@ -55,10 +63,22 @@ class Site {
                 $rawimage = file_get_contents($defaulticon);
             }
         } else {
-            $rawimage = file_get_contents($this->config->get('sitesdir').'/icons/'.$icon);
+            $rawimage = file_get_contents($this->config->get('sitesdir').'/icons/'.$this->iconname);
         }
-        $mimetype = (new \finfo(FILEINFO_MIME_TYPE))->buffer($rawimage);
-        return 'data:'.$mimetype.';base64,'.base64_encode($rawimage);
+        $imagedata = new stdClass();
+        $imagedata->mimetype = (new \finfo(FILEINFO_MIME_TYPE))->buffer($rawimage);
+        $imagedata->data = $rawimage;
+        return $imagedata;
+    }
+
+    /**
+     * Return a data uri or a site's favicon if an icon is not provided.
+     *
+     * @return string Base 64 encoded datauri for the icon image.
+     */
+    public function get_favicon_datauri(): string {
+        $imagedata = $this->get_favicon_image_data();
+        return 'data:'.$imagedata->mimetype.';base64,'.base64_encode($imagedata->data);
     }
 
 }